Another commit

+++ AdmissionWeb/src/main/by/epam/admissionweb/command/impl/admin/discipline/DeleteDisciplineCommand.java
@@ -69,18 +69,22 @@ public class DeleteDisciplineCommand implements Command {
 			String idStr = request.getParameter(CommandHelper.ParameterName.DISCIPLINE_ID);
 			int id = helper.parseId(idStr);
 			LOGGER.debug("COMMAND : DeleteDisciplineCommand (id = {})", id);
-			try {
-				DisciplineService service = ServiceFactory.getInstance().getDisciplineService();
-				boolean isDeleteEnable = service.isDeleteEnable();
-				if (!isDeleteEnable) {
-					request.setAttribute(CommandHelper.AttributeName.DELETE_UNABLE, true);
-				} else {
-					service.deleteDiscipline(id);
-					request.setAttribute(CommandHelper.AttributeName.SUCCESS_DELETED, true);
+			if (id == 0) {
+				request.setAttribute(CommandHelper.AttributeName.NOT_FOUND, true);
+			} else {
+				try {
+					DisciplineService service = ServiceFactory.getInstance().getDisciplineService();
+					boolean isDeleteEnable = service.isDeleteEnable();
+					if (!isDeleteEnable) {
+						request.setAttribute(CommandHelper.AttributeName.DELETE_UNABLE, true);
+					} else {
+						service.deleteDiscipline(id);
+						request.setAttribute(CommandHelper.AttributeName.SUCCESS_DELETED, true);
+					}
+				} catch (ServiceException ex) {
+					LOGGER.error(ex);
+					request.setAttribute(CommandHelper.AttributeName.ERROR, true);
 				}
-			} catch (ServiceException ex) {
-				LOGGER.error(ex);
-				request.setAttribute(CommandHelper.AttributeName.ERROR, true);
 			}
 			path = CommandHelper.PageName.EDIT_DISCIPLINE;
 		}
